[ API ] Refactoring Method and Ready to Modularity
